Add toggle for h2h v multi logging
