+++ app/javascript/packs/application.scss
@@ -3,6 +3,7 @@
 //Component Styles
 @import "../stylesheets/admin/users";
 @import "../stylesheets/sitewide/button";
+@import "../stylesheets/sitewide/toggle-switch";
 @import "../stylesheets/cards/card-profile";
 @import "../stylesheets/card-grid/card-grid";
 @import "../stylesheets/dashboard/dashboard";
